+++ routes/_app.tsx
@@ -1,0 +1,17 @@
+import { AppProps } from "$fresh/server.ts";
+import { Head } from "$fresh/runtime.ts";
+
+export default function App({ Component }: AppProps) {
+  return (
+    <>
+      <Head>
+        <title>Ana Components Docs</title>
+        <link
+          rel="stylesheet"
+          href="https://cdn.jsdelivr.net/gh/CarcajadaArtificial/ana-components@v0.0.13/static/styles.css"
+        />
+      </Head>
+      <Component />
+    </>
+  );
+}