+++ CHANGELOG.md
@@ -1,10 +1,21 @@
 # Changelog
 
-## v0.0.1
+## v0.0.2
 
 ### Added
 
-- Deno fresh setup
+- Navigation Island
+  - `~/islands/Navigation.tsx`
+
+- Text and input pages
+  - `~/routes/atoms/input.tsx`
+  - `~/routes/atoms/text.tsx`
+
+- App page
+  - `~/routes/_app.tsx`
+
+- Ana dependencies
+  `~/deps.ts`
 
 ### Comming soon (v1.0.0)